% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_p_values.R
\name{extract.p.values}
\alias{extract.p.values}
\title{Extract p-values}
\usage{
extract.p.values(manova.lists)
}
\arguments{
\item{manova.lists}{MANOVA list}
}
\description{
Extract p-values from MANOVA list
}

