% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_anova_test.R
\name{full.anova.test}
\alias{full.anova.test}
\title{ANOVA test for all landmarks}
\usage{
full.anova.test(group.1, group.2, iteracje = 1000, quant = 0.95)
}
\arguments{
\item{group.1}{first group}

\item{group.2}{second group}

\item{iteracje}{iterations}

\item{quant}{alpha}
}
\description{
Calculates anova test for all landmarks
}

