% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_dist.R
\name{land.dist}
\alias{land.dist}
\title{Distance between landmarks}
\usage{
land.dist(x, y)
}
\arguments{
\item{x}{shape 1}

\item{y}{shape 2}
}
\description{
Calculates distance between landmarks in two shapes.
}

