% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmark_anova.R
\name{landmark.anova}
\alias{landmark.anova}
\title{Landmark ANOVA}
\usage{
landmark.anova(group.1, group.2)
}
\arguments{
\item{group.1}{first group}

\item{group.2}{second group}
}
\description{
Calculates ANOVA for single landmark
}

