% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmark_anova_test.R
\name{landmark.anova.test}
\alias{landmark.anova.test}
\title{Landmark ANOVA test}
\usage{
landmark.anova.test(group.1, group.2, iteracje = 1000, quant = 0.95)
}
\arguments{
\item{group.1}{first group}

\item{group.2}{second group}

\item{iteracje}{iterations}

\item{quant}{alpha}
}
\description{
Calculates distribution for single landmark ANOVA
}

