% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarks_manova.R
\name{landmarks.manova}
\alias{landmarks.manova}
\title{MANOVA for single landmarks}
\usage{
landmarks.manova(data, takson = NULL, centroid.size = NULL)
}
\arguments{
\item{data}{data}

\item{takson}{taxon factor}

\item{centroid.size}{centroid size}
}
\description{
Calculate manova for single landmarks.
}

