% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_variance_ci.R
\name{landmarks.variance.ci}
\alias{landmarks.variance.ci}
\title{Confidence intervals of individual landmarks}
\usage{
landmarks.variance.ci(landmarks, iter = 1000, interval = 0.95)
}
\arguments{
\item{landmarks}{landmark array}

\item{iter}{number of iterations}

\item{interval}{confience interval}
}
\description{
Calculates confidence intervals of individual landmarks
}

