% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_variance.R
\name{landmarks.variance}
\alias{landmarks.variance}
\title{Individual variance of landmarks}
\usage{
landmarks.variance(landmarks)
}
\arguments{
\item{landmarks}{landmarks array}
}
\description{
Calculates individual variance of landmarks
}

