% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_variance_ci.R
\name{shape.variance.ci}
\alias{shape.variance.ci}
\title{Shape variance confidence intervals}
\usage{
shape.variance.ci(landmarks, iter = 1000, interval = 0.95)
}
\arguments{
\item{landmarks}{landmark array}

\item{iter}{number of iterations}

\item{interval}{confience interval}
}
\description{
Calculates shape variance confidence intervals
}

